requirements.txt
# Core ML Libraries
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2

# Time Series
statsmodels==0.14.1

# API and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
requests==2.31.0

# Data Quality and Profiling
pandas-profiling==3.6.6
great-expectations==0.18.8

# MLflow and DVC
mlflow==2.9.2
dvc[s3]>=3.0.0
dagshub==0.3.20

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-http==4.5.1

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# CI/CD (optional - only needed for GitHub Actions CML reports)
# cml>=0.0.1  # Uncomment if using CML in CI/CD (requires Python 3.11+)

# Storage
boto3==1.29.7
minio==7.2.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

